Refactor datahandlers into their own package
