+++ freqtrade/data/history/__init__.py
@@ -6,7 +6,7 @@ Includes:
 * download data from exchange and store to disk
 """
 # flake8: noqa: F401
+from .datahandlers import get_datahandler
 from .history_utils import (convert_trades_to_ohlcv, download_data_main, get_timerange, load_data,
                             load_pair_history, refresh_backtest_ohlcv_data,
                             refresh_backtest_trades_data, refresh_data, validate_backtest_data)
-from .idatahandler import get_datahandler
