+++ freqtrade/data/history/datahandlers/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa: F401
+from .idatahandler import IDataHandler, get_datahandler
