+++ freqtrade/data/history/datahandlers/hdf5datahandler.py
