+++ freqtrade/data/history/datahandlers/idatahandler.py
