+++ freqtrade/data/history/datahandlers/jsondatahandler.py
