+++ freqtrade/data/history/datahandlers/parquetdatahandler.py
